feat: refactoring to provide cross-lang support
