+++ src/injected/script.ts
@@ -19,13 +19,12 @@ import { Recorder } from './recorder';
 
 export default class Script {
   private _consoleAPI: ConsoleAPI  | undefined;
-  private _recorder: Recorder | undefined;
 
   constructor(injectedScript: InjectedScript, options: { enableRecorder: boolean }) {
     if ((window as any).playwright)
       return;
     this._consoleAPI = new ConsoleAPI(injectedScript);
     if (options.enableRecorder)
-      this._recorder = new Recorder(injectedScript, this._consoleAPI);
+       new Recorder(injectedScript, this._consoleAPI);
   }
 }
